Use 403 Forbidden instead of 404 Not Found for non-exposed API
Also: Just skip ApiExposedFilter for OptionsMethodProcessor (Jersey)

419045
