+++ org.eclipse.scout.rt.rest/src/main/java/org/eclipse/scout/rt/rest/resource/AbstractApiExposedFeature.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010, 2025 BSI Business Systems Integration AG
+ * Copyright (c) 2010, 2026 BSI Business Systems Integration AG
  *
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
@@ -73,6 +73,10 @@ public abstract class AbstractApiExposedFeature implements DynamicFeature {
 
   @Bean
   public interface IApiExposedFeatureFilterContributor {
+
+    /**
+     * @return {@code true} if this contributor handled this resource/feature (e.g. installed a more special filter or wants to skip filter registration)
+     */
     boolean configure(ResourceInfo resourceInfo, FeatureContext context);
   }
